requirements.txt
Authlib==1.3.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
Flask-Migrate
